Upgrade polkadart (#152)
* Upgrade polkadart & its dependencies in signer & tfchain, fix upgrade issue, produce new palletes, edit tfchain connect method

* Revert tfchain client changes, edit provider type

* Remove unused import & deprecated printTime

+++ packages/tfchain_client/test/contracts_test.dart
@@ -35,7 +35,7 @@ void main() {
 
     test('Test Get Active Contracts by wrong node Id', () async {
       try {
-        List<int> contracts =
+        List<BigInt> contracts =
             await setupManager.client.contracts.getActiveContracts(nodeId: -21);
         expect(contracts, isNotEmpty);
       } catch (error) {
